make python graphics generator and merge with font generator
