--- opendps/cv.h
@@ -1,34 +1,0 @@
-const uint8_t cv[] = {
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0xc3, 0x31, 0xa6, 0x18, 0xe3, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x21, 0x04, 0xc6, 0x58, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xbd, 0xf7, 0x84, 0x30,
-  0xff, 0xff, 0x21, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x04, 0xff, 0xff, 0x7b, 0xcf,
-  0x00, 0x00, 0xde, 0xfb, 0xf7, 0xbe, 0x63, 0x2c, 0x29, 0x65, 0x52, 0x8a, 0x73, 0xce, 0x52, 0xaa,
-  0xff, 0xff, 0x4a, 0x69, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x4a, 0x69, 0xff, 0xff, 0x42, 0x28,
-  0x52, 0x8a, 0xff, 0xff, 0x63, 0x4c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x04,
-  0xff, 0xff, 0x7b, 0xcf, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x7b, 0xcf, 0xff, 0xff, 0x10, 0x82,
-  0xad, 0x55, 0xf7, 0xde, 0x10, 0x82, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0xe7, 0x5c, 0xad, 0x75, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xad, 0x75, 0xde, 0xfb, 0x00, 0x00,
-  0xce, 0x59, 0xd6, 0xba, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0xad, 0x95, 0xdf, 0x1b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xdf, 0x1b, 0xa5, 0x34, 0x00, 0x00,
-  0xe7, 0x3c, 0xc6, 0x38, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x73, 0xae, 0xff, 0xff, 0x18, 0xc3, 0x00, 0x00, 0x18, 0xc3, 0xff, 0xff, 0x63, 0x2c, 0x00, 0x00,
-  0xf7, 0xde, 0xb5, 0xd6, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x31, 0xa6, 0xff, 0xff, 0x52, 0xaa, 0x00, 0x00, 0x52, 0xaa, 0xff, 0xff, 0x29, 0x45, 0x00, 0x00,
-  0xe7, 0x3c, 0xc6, 0x58, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x20, 0xef, 0x9d, 0x94, 0xb2, 0x00, 0x00, 0x94, 0xb2, 0xe7, 0x3c, 0x00, 0x00, 0x00, 0x00,
-  0xce, 0x79, 0xde, 0xfb, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0xad, 0x75, 0xce, 0x99, 0x00, 0x00, 0xce, 0x99, 0x9d, 0x13, 0x00, 0x00, 0x00, 0x00,
-  0xa5, 0x54, 0xff, 0xff, 0x18, 0xe3, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x6b, 0x4d, 0xff, 0xff, 0x29, 0x65, 0xff, 0xff, 0x5a, 0xeb, 0x00, 0x00, 0x00, 0x00,
-  0x4a, 0x49, 0xff, 0xff, 0x7c, 0x0f, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xb5, 0x96, 0xf7, 0xde, 0x10, 0xa2, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0xce, 0x99, 0xf7, 0xde, 0x7b, 0xef, 0x31, 0xa6, 0x42, 0x48, 0x8c, 0x71, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0xd6, 0xba, 0xff, 0xff, 0xc6, 0x58, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x10, 0x82, 0xb5, 0xd6, 0xf7, 0xde, 0xff, 0xff, 0xff, 0xff, 0xce, 0x79, 0x08, 0x41,
-  0x00, 0x00, 0x00, 0x00, 0x8c, 0x51, 0xff, 0xff, 0x7b, 0xef, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x61, 0x31, 0xc6, 0x18, 0xe3, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-};
-#define cv_width  16
-#define cv_height  15
