+++ opendps/font-large.c
@@ -1,11 +1,11 @@
 /** Font generated from gfx/Ubuntu-C.ttf 48pt */
 
-#include <stdint.h>
+#include "font-large.h"
 
 const uint32_t font_large_height = 35;
 const uint32_t font_large_num_glyphs = 13;
 
-const uint8_t font_large_0[] = {
+const uint8_t font_large_0[1120] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x61, 0x6b, 0x4d, 0xbe, 0x17, 0xe7, 0x3c, 
   0xf7, 0xbe, 0xd6, 0xda, 0xa5, 0x34, 0x42, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x24, 0xde, 0xfb, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
@@ -77,9 +77,8 @@ const uint8_t font_large_0[] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x61, 0x6b, 0x4d, 0xc6, 0x18, 0xe7, 0x3c, 
   0xf7, 0xbe, 0xd6, 0xda, 0xa5, 0x54, 0x42, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_large_0_len = 1120;
 
-const uint8_t font_large_1[] = {
+const uint8_t font_large_1[630] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x04, 0xf7, 0xbe, 
   0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xb5, 0xd6, 
@@ -121,9 +120,8 @@ const uint8_t font_large_1[] = {
   0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_large_1_len = 630;
 
-const uint8_t font_large_2[] = {
+const uint8_t font_large_2[1120] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x5a, 0xcb, 0xa5, 0x34, 0xd6, 0xda, 0xef, 0x9d, 0xf7, 0xbe, 
   0xdf, 0x1b, 0xb5, 0xb6, 0x6b, 0x8d, 0x08, 0x61, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x4a, 0x69, 0xd6, 0xda, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
@@ -195,9 +193,8 @@ const uint8_t font_large_2[] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_large_2_len = 1120;
 
-const uint8_t font_large_3[] = {
+const uint8_t font_large_3[1050] = {
   0x00, 0x00, 0x00, 0x00, 0x10, 0xa2, 0x73, 0x8e, 0xb5, 0xb6, 0xdf, 0x1b, 0xf7, 0xbe, 0xe7, 0x5c, 
   0xce, 0x59, 0x84, 0x30, 0x21, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 
   0x84, 0x30, 0xf7, 0xbe, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
@@ -265,9 +262,8 @@ const uint8_t font_large_3[] = {
   0x4a, 0x49, 0x94, 0xd2, 0xce, 0x79, 0xef, 0x7d, 0xf7, 0xde, 0xe7, 0x5c, 0xce, 0x79, 0x8c, 0x91, 
   0x39, 0xe7, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_large_3_len = 1050;
 
-const uint8_t font_large_4[] = {
+const uint8_t font_large_4[1190] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -344,9 +340,8 @@ const uint8_t font_large_4[] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_large_4_len = 1190;
 
-const uint8_t font_large_5[] = {
+const uint8_t font_large_5[980] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x7b, 0xef, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
@@ -410,9 +405,8 @@ const uint8_t font_large_5[] = {
   0xf7, 0xbe, 0xef, 0x9d, 0xd6, 0xda, 0xa5, 0x14, 0x52, 0xaa, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_large_5_len = 980;
 
-const uint8_t font_large_6[] = {
+const uint8_t font_large_6[1120] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -484,9 +478,8 @@ const uint8_t font_large_6[] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x31, 0x86, 0x94, 0xb2, 0xd6, 0xba, 0xef, 0x9d, 
   0xef, 0x7d, 0xd6, 0xba, 0x8c, 0x71, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_large_6_len = 1120;
 
-const uint8_t font_large_7[] = {
+const uint8_t font_large_7[1120] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0xd6, 0xba, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
@@ -558,9 +551,8 @@ const uint8_t font_large_7[] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_large_7_len = 1120;
 
-const uint8_t font_large_8[] = {
+const uint8_t font_large_8[1120] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x5a, 0xeb, 0xb5, 0x96, 0xe7, 0x3c, 
   0xf7, 0xbe, 0xe7, 0x3c, 0xb5, 0xb6, 0x63, 0x0c, 0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x29, 0x65, 0xd6, 0xba, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
@@ -632,9 +624,8 @@ const uint8_t font_large_8[] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0xe3, 0x7c, 0x0f, 0xc6, 0x58, 0xe7, 0x5c, 
   0xf7, 0xbe, 0xde, 0xfb, 0xa5, 0x14, 0x42, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_large_8_len = 1120;
 
-const uint8_t font_large_9[] = {
+const uint8_t font_large_9[1050] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0xa2, 0x7b, 0xef, 0xce, 0x59, 0xe7, 0x5c, 0xf7, 0xbe, 
   0xd6, 0xda, 0xa5, 0x34, 0x42, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x42, 0x28, 0xe7, 0x5c, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
@@ -702,9 +693,8 @@ const uint8_t font_large_9[] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_large_9_len = 1050;
 
-const uint8_t font_large_dot[] = {
+const uint8_t font_large_dot[350] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -728,9 +718,8 @@ const uint8_t font_large_dot[] = {
   0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xce, 0x99, 0xff, 0xff, 0xff, 0xff, 
   0xff, 0xff, 0xe7, 0x5c, 0x21, 0x04, 0xbd, 0xf7, 0xef, 0x9d, 0xc6, 0x58, 0x31, 0xc6
 };
-const uint32_t font_large_dot_len = 350;
 
-const uint8_t font_large_v[] = {
+const uint8_t font_large_v[1540] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xef, 0x7d, 0xff, 0xff, 
@@ -829,9 +818,8 @@ const uint8_t font_large_v[] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_large_v_len = 1540;
 
-const uint8_t font_large_a[] = {
+const uint8_t font_large_a[1540] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -930,7 +918,6 @@ const uint8_t font_large_a[] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_large_a_len = 1540;
 
 const uint8_t font_large_widths[13] = {
   16,
