+++ opendps/font-large.h
@@ -1,36 +1,30 @@
 /** Font generated from gfx/Ubuntu-C.ttf 48pt */
 
+#ifndef __FONT_LARGE_H__
+#define __FONT_LARGE_H__
+
+#include <stdint.h>
+
 #define FONT_LARGE_MAX_GLYPH_HEIGHT (35)
 #define FONT_LARGE_MAX_GLYPH_WIDTH  (22)
 
 extern const uint32_t font_large_height;
 extern const uint32_t font_large_num_glyphs;
-extern const uint8_t font_large_0[];
-extern const uint32_t font_large_0_len;
-extern const uint8_t font_large_1[];
-extern const uint32_t font_large_1_len;
-extern const uint8_t font_large_2[];
-extern const uint32_t font_large_2_len;
-extern const uint8_t font_large_3[];
-extern const uint32_t font_large_3_len;
-extern const uint8_t font_large_4[];
-extern const uint32_t font_large_4_len;
-extern const uint8_t font_large_5[];
-extern const uint32_t font_large_5_len;
-extern const uint8_t font_large_6[];
-extern const uint32_t font_large_6_len;
-extern const uint8_t font_large_7[];
-extern const uint32_t font_large_7_len;
-extern const uint8_t font_large_8[];
-extern const uint32_t font_large_8_len;
-extern const uint8_t font_large_9[];
-extern const uint32_t font_large_9_len;
-extern const uint8_t font_large_dot[];
-extern const uint32_t font_large_dot_len;
-extern const uint8_t font_large_v[];
-extern const uint32_t font_large_v_len;
-extern const uint8_t font_large_a[];
-extern const uint32_t font_large_a_len;
+extern const uint8_t font_large_0[1120];
+extern const uint8_t font_large_1[630];
+extern const uint8_t font_large_2[1120];
+extern const uint8_t font_large_3[1050];
+extern const uint8_t font_large_4[1190];
+extern const uint8_t font_large_5[980];
+extern const uint8_t font_large_6[1120];
+extern const uint8_t font_large_7[1120];
+extern const uint8_t font_large_8[1120];
+extern const uint8_t font_large_9[1050];
+extern const uint8_t font_large_dot[350];
+extern const uint8_t font_large_v[1540];
+extern const uint8_t font_large_a[1540];
 extern const uint8_t font_large_widths[13];
 extern const uint16_t font_large_sizes[13];
 extern const uint16_t *font_large_pix[13];
+
+#endif // __FONT_LARGE_H__
