+++ opendps/font-medium.c
@@ -1,11 +1,11 @@
 /** Font generated from gfx/Ubuntu-C.ttf 24pt */
 
-#include <stdint.h>
+#include "font-medium.h"
 
 const uint32_t font_medium_height = 17;
 const uint32_t font_medium_num_glyphs = 13;
 
-const uint8_t font_medium_0[] = {
+const uint8_t font_medium_0[272] = {
   0x00, 0x00, 0x00, 0x20, 0x8c, 0x91, 0xe7, 0x5c, 0xef, 0x9d, 0xb5, 0xd6, 0x21, 0x24, 0x00, 0x00, 
   0x00, 0x00, 0x94, 0xd2, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xde, 0xfb, 0x08, 0x41, 
   0x10, 0xa2, 0xf7, 0xde, 0xf7, 0xde, 0x42, 0x28, 0x18, 0xc3, 0xd6, 0xda, 0xff, 0xff, 0x63, 0x0c, 
@@ -24,9 +24,8 @@ const uint8_t font_medium_0[] = {
   0x00, 0x00, 0x94, 0xd2, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xde, 0xfb, 0x08, 0x41, 
   0x00, 0x00, 0x00, 0x20, 0x8c, 0x91, 0xe7, 0x5c, 0xef, 0x9d, 0xb5, 0xd6, 0x21, 0x24, 0x00, 0x00
 };
-const uint32_t font_medium_0_len = 272;
 
-const uint8_t font_medium_1[] = {
+const uint8_t font_medium_1[170] = {
   0x00, 0x00, 0x00, 0x00, 0x39, 0xe7, 0xf7, 0xde, 0xa5, 0x54, 0x00, 0x00, 0x18, 0xe3, 0xe7, 0x3c, 
   0xff, 0xff, 0xa5, 0x54, 0x21, 0x24, 0xde, 0xfb, 0xff, 0xff, 0xff, 0xff, 0xa5, 0x54, 0xef, 0x9d, 
   0xf7, 0xde, 0xb5, 0xb6, 0xff, 0xff, 0xa5, 0x54, 0xa5, 0x54, 0x42, 0x48, 0x63, 0x0c, 0xff, 0xff, 
@@ -39,9 +38,8 @@ const uint8_t font_medium_1[] = {
   0x63, 0x0c, 0xff, 0xff, 0xa5, 0x54, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xa5, 0x54, 
   0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xa5, 0x54
 };
-const uint32_t font_medium_1_len = 170;
 
-const uint8_t font_medium_2[] = {
+const uint8_t font_medium_2[272] = {
   0x10, 0x82, 0x8c, 0x71, 0xde, 0xfb, 0xf7, 0xbe, 0xdf, 0x1b, 0x94, 0xd2, 0x10, 0x82, 0x00, 0x00, 
   0xb5, 0xb6, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xce, 0x79, 0x00, 0x00, 
   0x3a, 0x07, 0xad, 0x55, 0x21, 0x24, 0x08, 0x41, 0x52, 0xaa, 0xf7, 0xde, 0xff, 0xff, 0x42, 0x48, 
@@ -60,9 +58,8 @@ const uint8_t font_medium_2[] = {
   0x7c, 0x0f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd6, 0xda, 
   0x8c, 0x71, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd6, 0xda
 };
-const uint32_t font_medium_2_len = 272;
 
-const uint8_t font_medium_3[] = {
+const uint8_t font_medium_3[272] = {
   0x21, 0x04, 0x9c, 0xf3, 0xe7, 0x3c, 0xf7, 0xbe, 0xd6, 0xba, 0x63, 0x4c, 0x00, 0x00, 0x00, 0x00, 
   0x9d, 0x13, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x6b, 0x8d, 0x00, 0x00, 
   0x29, 0x65, 0x7c, 0x0f, 0x21, 0x04, 0x08, 0x61, 0x8c, 0x71, 0xff, 0xff, 0xef, 0x7d, 0x00, 0x20, 
@@ -81,9 +78,8 @@ const uint8_t font_medium_3[] = {
   0xce, 0x99, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xad, 0x55, 0x00, 0x00, 
   0x42, 0x28, 0xb5, 0xb6, 0xef, 0x7d, 0xf7, 0xbe, 0xd6, 0xba, 0x7b, 0xef, 0x00, 0x20, 0x00, 0x00
 };
-const uint32_t font_medium_3_len = 272;
 
-const uint8_t font_medium_4[] = {
+const uint8_t font_medium_4[306] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x52, 0x8a, 0xff, 0xff, 0xff, 0xff, 0x08, 0x61, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xd6, 0xba, 0xff, 0xff, 0xff, 0xff, 
   0x08, 0x61, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x5a, 0xeb, 0xff, 0xff, 0xf7, 0xde, 
@@ -105,9 +101,8 @@ const uint8_t font_medium_4[] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xce, 0x99, 0xff, 0xff, 0x08, 0x61, 
   0x00, 0x00
 };
-const uint32_t font_medium_4_len = 306;
 
-const uint8_t font_medium_5[] = {
+const uint8_t font_medium_5[272] = {
   0x00, 0x00, 0x39, 0xe7, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x5a, 0xeb, 
   0x00, 0x00, 0x42, 0x28, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x5a, 0xeb, 
   0x00, 0x00, 0x4a, 0x89, 0xff, 0xff, 0x84, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -126,9 +121,8 @@ const uint8_t font_medium_5[] = {
   0x8c, 0x71, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf7, 0xde, 0x52, 0xaa, 0x00, 0x00, 
   0x4a, 0x89, 0xbe, 0x17, 0xef, 0x7d, 0xef, 0x9d, 0xb5, 0xd6, 0x39, 0xe7, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_medium_5_len = 272;
 
-const uint8_t font_medium_6[] = {
+const uint8_t font_medium_6[272] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x4a, 0x49, 0xb5, 0x96, 0xdf, 0x1b, 0x29, 0x65, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x9d, 0x13, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x4a, 0x69, 
   0x00, 0x00, 0x00, 0x00, 0x94, 0xb2, 0xff, 0xff, 0xdf, 0x1b, 0x6b, 0x4d, 0x29, 0x45, 0x00, 0x00, 
@@ -147,9 +141,8 @@ const uint8_t font_medium_6[] = {
   0x00, 0x00, 0x5a, 0xeb, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf7, 0xde, 0x42, 0x28, 
   0x00, 0x00, 0x00, 0x00, 0x6b, 0x6d, 0xd6, 0xda, 0xf7, 0xbe, 0xd6, 0xba, 0x5a, 0xcb, 0x00, 0x00
 };
-const uint32_t font_medium_6_len = 272;
 
-const uint8_t font_medium_7[] = {
+const uint8_t font_medium_7[272] = {
   0xe7, 0x5c, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
   0xe7, 0x5c, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x7b, 0xef, 0xff, 0xff, 0xad, 0x75, 
@@ -168,9 +161,8 @@ const uint8_t font_medium_7[] = {
   0x00, 0x00, 0x10, 0xa2, 0xff, 0xff, 0xf7, 0xde, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x29, 0x65, 0xff, 0xff, 0xe7, 0x3c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_medium_7_len = 272;
 
-const uint8_t font_medium_8[] = {
+const uint8_t font_medium_8[272] = {
   0x00, 0x00, 0x08, 0x61, 0x8c, 0x91, 0xd6, 0xda, 0xef, 0x7d, 0xbe, 0x17, 0x42, 0x28, 0x00, 0x00, 
   0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf7, 0xde, 0x3a, 0x07, 
   0x42, 0x08, 0xff, 0xff, 0xe7, 0x5c, 0x31, 0xa6, 0x10, 0x82, 0xad, 0x75, 0xff, 0xff, 0xb5, 0xb6, 
@@ -189,9 +181,8 @@ const uint8_t font_medium_8[] = {
   0x08, 0x61, 0xdf, 0x1b, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf7, 0xbe, 0x39, 0xc7, 
   0x00, 0x00, 0x10, 0xa2, 0x9d, 0x13, 0xef, 0x7d, 0xef, 0x9d, 0xb5, 0xd6, 0x31, 0xa6, 0x00, 0x00
 };
-const uint32_t font_medium_8_len = 272;
 
-const uint8_t font_medium_9[] = {
+const uint8_t font_medium_9[272] = {
   0x00, 0x00, 0x08, 0x61, 0x94, 0xd2, 0xe7, 0x5c, 0xef, 0x9d, 0xb5, 0xd6, 0x21, 0x44, 0x00, 0x00, 
   0x00, 0x00, 0xad, 0x55, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xdf, 0x1b, 0x08, 0x61, 
   0x21, 0x24, 0xff, 0xff, 0xf7, 0xbe, 0x42, 0x08, 0x18, 0xc3, 0xce, 0x79, 0xff, 0xff, 0x6b, 0x6d, 
@@ -210,9 +201,8 @@ const uint8_t font_medium_9[] = {
   0x00, 0x00, 0xd6, 0xda, 0xff, 0xff, 0xff, 0xff, 0xef, 0x7d, 0x4a, 0x49, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0xb5, 0x96, 0xce, 0x99, 0x84, 0x50, 0x18, 0xe3, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_medium_9_len = 272;
 
-const uint8_t font_medium_dot[] = {
+const uint8_t font_medium_dot[102] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -221,9 +211,8 @@ const uint8_t font_medium_dot[] = {
   0x00, 0x00, 0x00, 0x00, 0x39, 0xe7, 0xe7, 0x5c, 0xb5, 0xd6, 0x94, 0x92, 0xff, 0xff, 0xff, 0xff, 
   0x39, 0xe7, 0xe7, 0x5c, 0xb5, 0xd6
 };
-const uint32_t font_medium_dot_len = 102;
 
-const uint8_t font_medium_v[] = {
+const uint8_t font_medium_v[374] = {
   0x6b, 0x4d, 0xff, 0xff, 0xce, 0x79, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0xce, 0x99, 0xff, 0xff, 0x31, 0xa6, 0xff, 0xff, 0xf7, 0xbe, 0x00, 0x20, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0xf7, 0xde, 0xff, 0xff, 0x00, 0x20, 0xf7, 0xbe, 
@@ -249,9 +238,8 @@ const uint8_t font_medium_v[] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x52, 0xca, 0xff, 0xff, 0xff, 0xff, 0x42, 0x28, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_medium_v_len = 374;
 
-const uint8_t font_medium_a[] = {
+const uint8_t font_medium_a[408] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x39, 0xe7, 0xff, 0xff, 0xff, 0xff, 0x4a, 0x69, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x84, 0x30, 0xff, 0xff, 0xff, 0xff, 0x94, 0x92, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -279,7 +267,6 @@ const uint8_t font_medium_a[] = {
   0x63, 0x2c, 0xff, 0xff, 0xa5, 0x34, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0xa5, 0x54, 0xff, 0xff, 0x73, 0xce
 };
-const uint32_t font_medium_a_len = 408;
 
 const uint8_t font_medium_widths[13] = {
   8,
