+++ opendps/font-medium.h
@@ -1,36 +1,30 @@
 /** Font generated from gfx/Ubuntu-C.ttf 24pt */
 
+#ifndef __FONT_MEDIUM_H__
+#define __FONT_MEDIUM_H__
+
+#include <stdint.h>
+
 #define FONT_MEDIUM_MAX_GLYPH_HEIGHT (17)
 #define FONT_MEDIUM_MAX_GLYPH_WIDTH  (12)
 
 extern const uint32_t font_medium_height;
 extern const uint32_t font_medium_num_glyphs;
-extern const uint8_t font_medium_0[];
-extern const uint32_t font_medium_0_len;
-extern const uint8_t font_medium_1[];
-extern const uint32_t font_medium_1_len;
-extern const uint8_t font_medium_2[];
-extern const uint32_t font_medium_2_len;
-extern const uint8_t font_medium_3[];
-extern const uint32_t font_medium_3_len;
-extern const uint8_t font_medium_4[];
-extern const uint32_t font_medium_4_len;
-extern const uint8_t font_medium_5[];
-extern const uint32_t font_medium_5_len;
-extern const uint8_t font_medium_6[];
-extern const uint32_t font_medium_6_len;
-extern const uint8_t font_medium_7[];
-extern const uint32_t font_medium_7_len;
-extern const uint8_t font_medium_8[];
-extern const uint32_t font_medium_8_len;
-extern const uint8_t font_medium_9[];
-extern const uint32_t font_medium_9_len;
-extern const uint8_t font_medium_dot[];
-extern const uint32_t font_medium_dot_len;
-extern const uint8_t font_medium_v[];
-extern const uint32_t font_medium_v_len;
-extern const uint8_t font_medium_a[];
-extern const uint32_t font_medium_a_len;
+extern const uint8_t font_medium_0[272];
+extern const uint8_t font_medium_1[170];
+extern const uint8_t font_medium_2[272];
+extern const uint8_t font_medium_3[272];
+extern const uint8_t font_medium_4[306];
+extern const uint8_t font_medium_5[272];
+extern const uint8_t font_medium_6[272];
+extern const uint8_t font_medium_7[272];
+extern const uint8_t font_medium_8[272];
+extern const uint8_t font_medium_9[272];
+extern const uint8_t font_medium_dot[102];
+extern const uint8_t font_medium_v[374];
+extern const uint8_t font_medium_a[408];
 extern const uint8_t font_medium_widths[13];
 extern const uint16_t font_medium_sizes[13];
 extern const uint16_t *font_medium_pix[13];
+
+#endif // __FONT_MEDIUM_H__
