+++ opendps/font-small.c
@@ -1,11 +1,11 @@
 /** Font generated from gfx/Ubuntu-C.ttf 18pt */
 
-#include <stdint.h>
+#include "font-small.h"
 
 const uint32_t font_small_height = 12;
 const uint32_t font_small_num_glyphs = 13;
 
-const uint8_t font_small_0[] = {
+const uint8_t font_small_0[144] = {
   0x00, 0x20, 0x9c, 0xd3, 0xef, 0x9d, 0xe7, 0x3c, 0x63, 0x2c, 0x00, 0x00, 0x6b, 0x6d, 0xf7, 0xde, 
   0x31, 0x86, 0x6b, 0x4d, 0xf7, 0xde, 0x29, 0x45, 0xbe, 0x17, 0xc6, 0x18, 0x00, 0x00, 0x08, 0x41, 
   0xf7, 0xde, 0x7b, 0xef, 0xef, 0x7d, 0x9d, 0x13, 0x00, 0x00, 0x00, 0x00, 0xe7, 0x3c, 0xa5, 0x34, 
@@ -16,9 +16,8 @@ const uint8_t font_small_0[] = {
   0x00, 0x00, 0x08, 0x41, 0xf7, 0xde, 0x7b, 0xef, 0x6b, 0x6d, 0xf7, 0xbe, 0x29, 0x65, 0x63, 0x2c, 
   0xf7, 0xde, 0x29, 0x65, 0x00, 0x20, 0x9c, 0xf3, 0xef, 0x9d, 0xe7, 0x3c, 0x6b, 0x4d, 0x00, 0x00
 };
-const uint32_t font_small_0_len = 144;
 
-const uint8_t font_small_1[] = {
+const uint8_t font_small_1[96] = {
   0x00, 0x00, 0x00, 0x00, 0x6b, 0x6d, 0xff, 0xff, 0x00, 0x00, 0x18, 0xc3, 0xe7, 0x5c, 0xff, 0xff, 
   0x00, 0x20, 0xbd, 0xf7, 0xd6, 0xda, 0xff, 0xff, 0x6b, 0x8d, 0x73, 0x8e, 0x8c, 0x51, 0xff, 0xff, 
   0x00, 0x00, 0x00, 0x00, 0x8c, 0x51, 0xff, 0xff, 0x00, 0x00, 0x00, 0x00, 0x8c, 0x51, 0xff, 0xff, 
@@ -26,9 +25,8 @@ const uint8_t font_small_1[] = {
   0x00, 0x00, 0x00, 0x00, 0x8c, 0x51, 0xff, 0xff, 0x00, 0x00, 0x00, 0x00, 0x8c, 0x51, 0xff, 0xff, 
   0x00, 0x00, 0x00, 0x00, 0x8c, 0x51, 0xff, 0xff, 0x00, 0x00, 0x00, 0x00, 0x8c, 0x51, 0xff, 0xff
 };
-const uint32_t font_small_1_len = 96;
 
-const uint8_t font_small_2[] = {
+const uint8_t font_small_2[144] = {
   0x73, 0x8e, 0xde, 0xfb, 0xf7, 0xbe, 0xce, 0x79, 0x42, 0x28, 0x00, 0x00, 0x84, 0x30, 0x29, 0x65, 
   0x08, 0x61, 0x9c, 0xf3, 0xef, 0x9d, 0x10, 0xa2, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x42, 0x28, 
   0xff, 0xff, 0x4a, 0x69, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x52, 0xaa, 0xff, 0xff, 0x4a, 0x69, 
@@ -39,9 +37,8 @@ const uint8_t font_small_2[] = {
   0x21, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xd6, 0x9a, 0xd6, 0xba, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0xe7, 0x5c, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x9d, 0x13
 };
-const uint32_t font_small_2_len = 144;
 
-const uint8_t font_small_3[] = {
+const uint8_t font_small_3[144] = {
   0x7b, 0xef, 0xd6, 0xda, 0xef, 0x7d, 0xb5, 0xd6, 0x21, 0x24, 0x00, 0x00, 0x63, 0x2c, 0x21, 0x04, 
   0x10, 0xa2, 0xc6, 0x58, 0xce, 0x99, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x6b, 0x4d, 
   0xff, 0xff, 0x18, 0xc3, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x73, 0xce, 0xff, 0xff, 0x10, 0xa2, 
@@ -52,9 +49,8 @@ const uint8_t font_small_3[] = {
   0x00, 0x00, 0x29, 0x65, 0xff, 0xff, 0x5a, 0xeb, 0x63, 0x2c, 0x10, 0xa2, 0x10, 0xa2, 0xad, 0x95, 
   0xe7, 0x5c, 0x10, 0x82, 0x9d, 0x13, 0xe7, 0x3c, 0xef, 0x7d, 0xbd, 0xf7, 0x31, 0x86, 0x00, 0x00
 };
-const uint32_t font_small_3_len = 144;
 
-const uint8_t font_small_4[] = {
+const uint8_t font_small_4[144] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xce, 0x79, 0xff, 0xff, 0x08, 0x41, 0x00, 0x00, 0x00, 0x00, 
   0x5a, 0xeb, 0xff, 0xff, 0xff, 0xff, 0x08, 0x41, 0x00, 0x00, 0x00, 0x00, 0xd6, 0xda, 0xc6, 0x38, 
   0xff, 0xff, 0x08, 0x41, 0x00, 0x00, 0x5a, 0xeb, 0xdf, 0x1b, 0x63, 0x2c, 0xff, 0xff, 0x08, 0x41, 
@@ -65,9 +61,8 @@ const uint8_t font_small_4[] = {
   0x00, 0x00, 0x5a, 0xeb, 0xff, 0xff, 0x08, 0x41, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x5a, 0xeb, 
   0xff, 0xff, 0x08, 0x41, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x5a, 0xeb, 0xff, 0xff, 0x08, 0x41
 };
-const uint32_t font_small_4_len = 144;
 
-const uint8_t font_small_5[] = {
+const uint8_t font_small_5[144] = {
   0x00, 0x00, 0xef, 0x9d, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x42, 0x28, 0x00, 0x00, 0xf7, 0xde, 
   0x63, 0x2c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0xff, 0xff, 0x52, 0xaa, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x10, 0x82, 0xff, 0xff, 0x4a, 0x49, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -78,9 +73,8 @@ const uint8_t font_small_5[] = {
   0x00, 0x00, 0x29, 0x65, 0xff, 0xff, 0x5a, 0xeb, 0x39, 0xc7, 0x08, 0x61, 0x18, 0xe3, 0xbd, 0xf7, 
   0xd6, 0xba, 0x08, 0x41, 0xad, 0x55, 0xef, 0x7d, 0xe7, 0x3c, 0xa5, 0x34, 0x18, 0xe3, 0x00, 0x00
 };
-const uint32_t font_small_5_len = 144;
 
-const uint8_t font_small_6[] = {
+const uint8_t font_small_6[144] = {
   0x00, 0x00, 0x00, 0x00, 0x29, 0x65, 0xa5, 0x54, 0xe7, 0x3c, 0x29, 0x65, 0x00, 0x00, 0x39, 0xe7, 
   0xef, 0x9d, 0xad, 0x75, 0x31, 0x86, 0x00, 0x00, 0x00, 0x20, 0xdf, 0x1b, 0xad, 0x75, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x5a, 0xcb, 0xff, 0xff, 0x31, 0xa6, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -91,9 +85,8 @@ const uint8_t font_small_6[] = {
   0x00, 0x00, 0x00, 0x00, 0xbd, 0xf7, 0xc6, 0x38, 0x42, 0x28, 0xff, 0xff, 0x4a, 0x69, 0x29, 0x45, 
   0xf7, 0xbe, 0x73, 0xae, 0x00, 0x00, 0x73, 0xce, 0xe7, 0x5c, 0xe7, 0x5c, 0x94, 0xd2, 0x00, 0x20
 };
-const uint32_t font_small_6_len = 144;
 
-const uint8_t font_small_7[] = {
+const uint8_t font_small_7[144] = {
   0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xc6, 0x38, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x29, 0x45, 0xf7, 0xde, 0x6b, 0x6d, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x9c, 0xf3, 
   0xe7, 0x5c, 0x08, 0x41, 0x00, 0x00, 0x00, 0x00, 0x08, 0x41, 0xef, 0x9d, 0x84, 0x50, 0x00, 0x00, 
@@ -104,9 +97,8 @@ const uint8_t font_small_7[] = {
   0xef, 0x9d, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xc6, 0x38, 0xd6, 0x9a, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xdf, 0x1b, 0xb5, 0xb6, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_small_7_len = 144;
 
-const uint8_t font_small_8[] = {
+const uint8_t font_small_8[144] = {
   0x00, 0x20, 0x94, 0xd2, 0xe7, 0x5c, 0xe7, 0x3c, 0x84, 0x30, 0x00, 0x00, 0x84, 0x50, 0xdf, 0x1b, 
   0x18, 0xe3, 0x39, 0xe7, 0xf7, 0xde, 0x63, 0x2c, 0xce, 0x79, 0xa5, 0x34, 0x00, 0x00, 0x00, 0x00, 
   0xd6, 0xda, 0xa5, 0x34, 0xce, 0x59, 0xb5, 0xd6, 0x00, 0x00, 0x00, 0x00, 0xe7, 0x3c, 0x94, 0xd2, 
@@ -117,9 +109,8 @@ const uint8_t font_small_8[] = {
   0x00, 0x00, 0x00, 0x00, 0xad, 0x95, 0xc6, 0x58, 0xbd, 0xd7, 0xce, 0x99, 0x18, 0xc3, 0x39, 0xc7, 
   0xf7, 0xbe, 0x63, 0x2c, 0x10, 0xa2, 0xa5, 0x54, 0xef, 0x7d, 0xdf, 0x1b, 0x73, 0xae, 0x00, 0x00
 };
-const uint32_t font_small_8_len = 144;
 
-const uint8_t font_small_9[] = {
+const uint8_t font_small_9[144] = {
   0x08, 0x41, 0x9d, 0x13, 0xe7, 0x5c, 0xe7, 0x3c, 0x6b, 0x4d, 0x00, 0x00, 0x84, 0x30, 0xef, 0x7d, 
   0x21, 0x24, 0x5a, 0xeb, 0xff, 0xff, 0x31, 0x86, 0xd6, 0xba, 0xad, 0x75, 0x00, 0x00, 0x00, 0x00, 
   0xef, 0x7d, 0x84, 0x30, 0xef, 0x7d, 0x94, 0xd2, 0x00, 0x00, 0x00, 0x00, 0xce, 0x79, 0xad, 0x95, 
@@ -130,16 +121,14 @@ const uint8_t font_small_9[] = {
   0x00, 0x00, 0xc6, 0x38, 0xd6, 0xba, 0x00, 0x00, 0x00, 0x20, 0x39, 0xc7, 0xbd, 0xf7, 0xef, 0x7d, 
   0x29, 0x85, 0x00, 0x00, 0x5a, 0xcb, 0xdf, 0x1b, 0x9c, 0xf3, 0x21, 0x04, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_small_9_len = 144;
 
-const uint8_t font_small_dot[] = {
+const uint8_t font_small_dot[48] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xb5, 0xb6, 0xd6, 0xda, 0xb5, 0xd6, 0xd6, 0xda
 };
-const uint32_t font_small_dot_len = 48;
 
-const uint8_t font_small_v[] = {
+const uint8_t font_small_v[216] = {
   0x84, 0x30, 0xff, 0xff, 0x21, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 
   0x73, 0xce, 0x4a, 0x89, 0xff, 0xff, 0x52, 0xaa, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x52, 0xaa, 
   0xff, 0xff, 0x42, 0x08, 0x18, 0xc3, 0xff, 0xff, 0x84, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
@@ -155,9 +144,8 @@ const uint8_t font_small_v[] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x8c, 0x71, 0xff, 0xff, 
   0x7b, 0xef, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 };
-const uint32_t font_small_v_len = 216;
 
-const uint8_t font_small_a[] = {
+const uint8_t font_small_a[216] = {
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x84, 0x50, 0xff, 0xff, 0x94, 0x92, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xe7, 0x3c, 0xf7, 0xde, 0xef, 0x7d, 0x00, 0x20, 
   0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x39, 0xe7, 0xff, 0xff, 0x84, 0x10, 0xff, 0xff, 
@@ -173,7 +161,6 @@ const uint8_t font_small_a[] = {
   0x31, 0xa6, 0xff, 0xff, 0x52, 0xaa, 0x84, 0x30, 0xf7, 0xde, 0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 
   0x00, 0x00, 0x00, 0x20, 0xf7, 0xde, 0x8c, 0x91
 };
-const uint32_t font_small_a_len = 216;
 
 const uint8_t font_small_widths[13] = {
   6,
