+++ opendps/font-small.h
@@ -1,36 +1,30 @@
 /** Font generated from gfx/Ubuntu-C.ttf 18pt */
 
+#ifndef __FONT_SMALL_H__
+#define __FONT_SMALL_H__
+
+#include <stdint.h>
+
 #define FONT_SMALL_MAX_GLYPH_HEIGHT (12)
 #define FONT_SMALL_MAX_GLYPH_WIDTH  (9)
 
 extern const uint32_t font_small_height;
 extern const uint32_t font_small_num_glyphs;
-extern const uint8_t font_small_0[];
-extern const uint32_t font_small_0_len;
-extern const uint8_t font_small_1[];
-extern const uint32_t font_small_1_len;
-extern const uint8_t font_small_2[];
-extern const uint32_t font_small_2_len;
-extern const uint8_t font_small_3[];
-extern const uint32_t font_small_3_len;
-extern const uint8_t font_small_4[];
-extern const uint32_t font_small_4_len;
-extern const uint8_t font_small_5[];
-extern const uint32_t font_small_5_len;
-extern const uint8_t font_small_6[];
-extern const uint32_t font_small_6_len;
-extern const uint8_t font_small_7[];
-extern const uint32_t font_small_7_len;
-extern const uint8_t font_small_8[];
-extern const uint32_t font_small_8_len;
-extern const uint8_t font_small_9[];
-extern const uint32_t font_small_9_len;
-extern const uint8_t font_small_dot[];
-extern const uint32_t font_small_dot_len;
-extern const uint8_t font_small_v[];
-extern const uint32_t font_small_v_len;
-extern const uint8_t font_small_a[];
-extern const uint32_t font_small_a_len;
+extern const uint8_t font_small_0[144];
+extern const uint8_t font_small_1[96];
+extern const uint8_t font_small_2[144];
+extern const uint8_t font_small_3[144];
+extern const uint8_t font_small_4[144];
+extern const uint8_t font_small_5[144];
+extern const uint8_t font_small_6[144];
+extern const uint8_t font_small_7[144];
+extern const uint8_t font_small_8[144];
+extern const uint8_t font_small_9[144];
+extern const uint8_t font_small_dot[48];
+extern const uint8_t font_small_v[216];
+extern const uint8_t font_small_a[216];
 extern const uint8_t font_small_widths[13];
 extern const uint16_t font_small_sizes[13];
 extern const uint16_t *font_small_pix[13];
+
+#endif // __FONT_SMALL_H__
