+++ opendps/gfx-cc.h
@@ -1,0 +1,13 @@
+/** Gfx generated from gfx\png\cc.png */
+
+#ifndef __GFX_CC_H__
+#define __GFX_CC_H__
+
+#include <stdint.h>
+
+#define GFX_CC_HEIGHT (15)
+#define GFX_CC_WIDTH  (16)
+
+extern const uint8_t gfx_cc[480];
+
+#endif // __GFX_CC_H__
