+++ opendps/gfx-cv.h
@@ -1,0 +1,13 @@
+/** Gfx generated from gfx\png\cv.png */
+
+#ifndef __GFX_CV_H__
+#define __GFX_CV_H__
+
+#include <stdint.h>
+
+#define GFX_CV_HEIGHT (15)
+#define GFX_CV_WIDTH  (16)
+
+extern const uint8_t gfx_cv[480];
+
+#endif // __GFX_CV_H__
