+++ opendps/gfx-padlock.c
@@ -1,0 +1,30 @@
+/** Gfx generated from gfx\png\padlock.png */
+
+#include "gfx-padlock.h"
+
+const uint8_t gfx_padlock[384] = {
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x42, 0x28, 0x9c, 0xf3, 0x9c, 0xf3, 0x42, 0x28, 
+  0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x24, 0xbe, 0x17, 
+  0xff, 0xff, 0xf7, 0xbe, 0xf7, 0xbe, 0xff, 0xff, 0xbe, 0x17, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x10, 0x82, 0xce, 0x99, 0xef, 0x7d, 0x63, 0x0c, 0x18, 0xe3, 0x18, 0xe3, 0x63, 0x0c, 
+  0xef, 0x7d, 0xce, 0x99, 0x10, 0x82, 0x00, 0x00, 0x00, 0x00, 0x6b, 0x6d, 0xff, 0xff, 0x52, 0x8a, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x52, 0x8a, 0xff, 0xff, 0x6b, 0x6d, 0x00, 0x00, 
+  0x00, 0x00, 0xad, 0x75, 0xdf, 0x1b, 0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x20, 0xdf, 0x1b, 0xad, 0x75, 0x00, 0x00, 0x00, 0x00, 0xb5, 0xd6, 0xd6, 0xba, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xd6, 0xba, 0xb5, 0xd6, 0x00, 0x00, 
+  0x00, 0x00, 0xb5, 0xb6, 0xd6, 0xba, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0xd6, 0xba, 0xb5, 0xb6, 0x00, 0x00, 0xb5, 0xd6, 0xe7, 0x5c, 0xef, 0x9d, 0xb5, 0xd6, 
+  0xb5, 0xd6, 0xb5, 0xd6, 0xb5, 0xd6, 0xb5, 0xd6, 0xb5, 0xd6, 0xef, 0x9d, 0xe7, 0x5c, 0xb5, 0xd6, 
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff
+};
