+++ opendps/gfx-padlock.h
@@ -1,0 +1,13 @@
+/** Gfx generated from gfx\png\padlock.png */
+
+#ifndef __GFX_PADLOCK_H__
+#define __GFX_PADLOCK_H__
+
+#include <stdint.h>
+
+#define GFX_PADLOCK_HEIGHT (16)
+#define GFX_PADLOCK_WIDTH  (12)
+
+extern const uint8_t gfx_padlock[384];
+
+#endif // __GFX_PADLOCK_H__
