+++ opendps/gfx-power.h
@@ -1,0 +1,13 @@
+/** Gfx generated from gfx\png\power.png */
+
+#ifndef __GFX_POWER_H__
+#define __GFX_POWER_H__
+
+#include <stdint.h>
+
+#define GFX_POWER_HEIGHT (16)
+#define GFX_POWER_WIDTH  (16)
+
+extern const uint8_t gfx_power[512];
+
+#endif // __GFX_POWER_H__
