+++ opendps/gfx-thermometer.h
@@ -1,0 +1,13 @@
+/** Gfx generated from gfx\png\thermometer.png */
+
+#ifndef __GFX_THERMOMETER_H__
+#define __GFX_THERMOMETER_H__
+
+#include <stdint.h>
+
+#define GFX_THERMOMETER_HEIGHT (64)
+#define GFX_THERMOMETER_WIDTH  (26)
+
+extern const uint8_t gfx_thermometer[3328];
+
+#endif // __GFX_THERMOMETER_H__
