+++ opendps/gfx-wifi.c
@@ -1,0 +1,44 @@
+/** Gfx generated from gfx\png\wifi.png */
+
+#include "gfx-wifi.h"
+
+const uint8_t gfx_wifi[608] = {
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x29, 0x65, 0x52, 0xca, 
+  0x73, 0xae, 0xbd, 0xd7, 0x7b, 0xcf, 0x5a, 0xcb, 0x31, 0x86, 0x08, 0x41, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0xe3, 0x7b, 0xcf, 
+  0xce, 0x59, 0xf7, 0xbe, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf7, 0xbe, 
+  0xce, 0x99, 0x7c, 0x0f, 0x21, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 
+  0x6b, 0x6d, 0xdf, 0x1b, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe7, 0x3c, 0xce, 0x59, 0xbe, 0x17, 
+  0xce, 0x59, 0xe7, 0x3c, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe7, 0x3c, 0x73, 0xae, 0x00, 0x20, 
+  0x00, 0x00, 0x18, 0xc3, 0xa5, 0x54, 0xff, 0xff, 0xff, 0xff, 0xe7, 0x5c, 0x8c, 0x71, 0x39, 0xc7, 
+  0x10, 0xa2, 0x10, 0xa2, 0x18, 0xc3, 0x10, 0xa2, 0x10, 0xa2, 0x39, 0xc7, 0x8c, 0x71, 0xe7, 0x3c, 
+  0xff, 0xff, 0xff, 0xff, 0xb5, 0xb6, 0x18, 0xe3, 0x31, 0xa6, 0xef, 0x7d, 0xff, 0xff, 0x9d, 0x13, 
+  0x21, 0x04, 0x21, 0x24, 0x7b, 0xcf, 0xbd, 0xf7, 0xde, 0xfb, 0xe7, 0x3c, 0xde, 0xfb, 0xbd, 0xf7, 
+  0x7b, 0xef, 0x21, 0x44, 0x21, 0x04, 0x9d, 0x13, 0xff, 0xff, 0xe7, 0x3c, 0x31, 0x86, 0x00, 0x00, 
+  0x42, 0x28, 0x6b, 0x6d, 0x10, 0x82, 0x84, 0x30, 0xef, 0x7d, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xef, 0x9d, 0x8c, 0x71, 0x10, 0xa2, 0x6b, 0x4d, 
+  0x39, 0xc7, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xa5, 0x54, 0xff, 0xff, 0xff, 0xff, 
+  0xde, 0xfb, 0x8c, 0x71, 0x52, 0xaa, 0x42, 0x28, 0x52, 0xaa, 0x8c, 0x71, 0xd6, 0xda, 0xff, 0xff, 
+  0xff, 0xff, 0xad, 0x75, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x42, 0x48, 0xde, 0xfb, 0x8c, 0x51, 0x10, 0xa2, 0x29, 0x65, 0x63, 0x4c, 0x7b, 0xef, 0x6b, 0x4d, 
+  0x29, 0x85, 0x10, 0xa2, 0x84, 0x30, 0xd6, 0xda, 0x3a, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0xc3, 0x21, 0x04, 0xa5, 0x54, 0xf7, 0xbe, 
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf7, 0xde, 0xa5, 0x54, 0x21, 0x04, 0x10, 0xa2, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x39, 0xe7, 0xef, 0x9d, 0xff, 0xff, 0xc6, 0x18, 0xa5, 0x34, 0xc6, 0x18, 0xff, 0xff, 0xef, 0x7d, 
+  0x31, 0xa6, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x42, 0x08, 0x4a, 0x69, 0x08, 0x61, 0x21, 0x04, 
+  0x08, 0x61, 0x4a, 0x69, 0x39, 0xe7, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x21, 0x24, 0xce, 0x59, 0xef, 0x9d, 0xce, 0x59, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xa5, 0x54, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xa5, 0x54, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 
+  0xff, 0xff, 0xff, 0xff, 0xbd, 0xf7, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x5a, 0xeb, 0xf7, 0xde, 0xff, 0xff, 0xf7, 0xde, 0x5a, 0xeb, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x42, 0x28, 0x94, 0xb2, 0x42, 0x28, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+};
