+++ opendps/gfx-wifi.h
@@ -1,0 +1,13 @@
+/** Gfx generated from gfx\png\wifi.png */
+
+#ifndef __GFX_WIFI_H__
+#define __GFX_WIFI_H__
+
+#include <stdint.h>
+
+#define GFX_WIFI_HEIGHT (16)
+#define GFX_WIFI_WIDTH  (19)
+
+extern const uint8_t gfx_wifi[608];
+
+#endif // __GFX_WIFI_H__
