--- opendps/power.h
@@ -1,36 +1,0 @@
-const uint8_t power[] = {
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x4c,
-  0x63, 0x2c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0xe3, 0xf7, 0xbe,
-  0xef, 0x9d, 0x18, 0xc3, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x41, 0x00, 0x00, 0x00, 0x00, 0x21, 0x04, 0xf7, 0xbe,
-  0xef, 0x9d, 0x18, 0xe3, 0x00, 0x00, 0x00, 0x00, 0x08, 0x41, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x4a, 0x89, 0xd6, 0xda, 0x8c, 0x71, 0x00, 0x00, 0x21, 0x04, 0xf7, 0xbe,
-  0xef, 0x9d, 0x18, 0xe3, 0x00, 0x00, 0x8c, 0x91, 0xd6, 0xda, 0x4a, 0x69, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x31, 0xa6, 0xef, 0x9d, 0xff, 0xff, 0x94, 0xb2, 0x00, 0x00, 0x21, 0x04, 0xf7, 0xbe,
-  0xef, 0x9d, 0x18, 0xe3, 0x00, 0x00, 0x9c, 0xd3, 0xff, 0xff, 0xef, 0x7d, 0x31, 0x86, 0x00, 0x00,
-  0x00, 0x20, 0xbd, 0xf7, 0xff, 0xff, 0xad, 0x75, 0x08, 0x41, 0x00, 0x00, 0x21, 0x04, 0xf7, 0xbe,
-  0xef, 0x9d, 0x18, 0xe3, 0x00, 0x00, 0x08, 0x41, 0xb5, 0x96, 0xff, 0xff, 0xb5, 0xd6, 0x00, 0x00,
-  0x31, 0xa6, 0xf7, 0xde, 0xef, 0x7d, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x18, 0xe3, 0xf7, 0xde,
-  0xf7, 0xbe, 0x18, 0xc3, 0x00, 0x00, 0x00, 0x00, 0x29, 0x45, 0xef, 0x9d, 0xf7, 0xde, 0x31, 0x86,
-  0x7b, 0xcf, 0xff, 0xff, 0xb5, 0xd6, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x5a, 0xcb,
-  0x52, 0xaa, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0x63, 0x4c,
-  0xdf, 0x1b, 0xff, 0xff, 0x9c, 0xf3, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xa5, 0x34, 0xff, 0xff, 0x84, 0x30,
-  0x7b, 0xcf, 0xff, 0xff, 0xb5, 0xd6, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0x63, 0x4c,
-  0x31, 0xa6, 0xf7, 0xde, 0xef, 0x7d, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x29, 0x45, 0xef, 0x9d, 0xf7, 0xde, 0x31, 0x86,
-  0x00, 0x20, 0xbd, 0xf7, 0xff, 0xff, 0xad, 0x75, 0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0xb5, 0xb6, 0xff, 0xff, 0xb5, 0xd6, 0x00, 0x00,
-  0x00, 0x00, 0x31, 0xa6, 0xef, 0x7d, 0xff, 0xff, 0xad, 0x75, 0x21, 0x44, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x29, 0x45, 0xb5, 0x96, 0xff, 0xff, 0xef, 0x7d, 0x31, 0x86, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x4a, 0x69, 0xe7, 0x5c, 0xff, 0xff, 0xef, 0x9d, 0xad, 0x75, 0x84, 0x30,
-  0x84, 0x30, 0xad, 0x95, 0xef, 0x9d, 0xff, 0xff, 0xe7, 0x5c, 0x4a, 0x49, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x29, 0x65, 0xa5, 0x54, 0xef, 0x9d, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xef, 0x9d, 0xa5, 0x34, 0x29, 0x45, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x24, 0x5a, 0xcb, 0x94, 0xb2,
-  0x94, 0xb2, 0x5a, 0xcb, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-};
-#define power_width  16
-#define power_height  16
