--- opendps/thermometer.h
@@ -1,212 +1,0 @@
-const uint8_t thermometer[] = {
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x18, 0xe3, 0x84, 0x10, 0xce, 0x79, 0xf7, 0xde, 0xf7, 0xbe, 0xbd, 0xf7, 0x6b, 0x6d,
-  0x08, 0x61, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x6b, 0x6d, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xf7, 0xbe, 0x4a, 0x49, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x7c, 0x0f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xe7, 0x5c, 0xe7, 0x5c, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x52, 0xaa, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x42, 0x28, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xbd, 0xf7, 0x29, 0x85, 0x00, 0x00, 0x00, 0x00, 0x42, 0x28, 0xd6, 0xba, 0xff, 0xff, 0xff, 0xff,
-  0xf7, 0xde, 0x18, 0xe3, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xce, 0x99, 0xff, 0xff,
-  0xff, 0xff, 0x9c, 0xf3, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0xce, 0x59, 0xff, 0xff, 0xff, 0xff, 0x94, 0xb2, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x29, 0x65,
-  0xff, 0xff, 0xff, 0xff, 0xef, 0x7d, 0x10, 0x82, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x31, 0xa6, 0xff, 0xff, 0xff, 0xff, 0xd6, 0xda, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x52, 0x8a, 0xff, 0xff, 0xff, 0xff, 0x94, 0xd2, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xce, 0x99, 0xff, 0xff, 0xf7, 0xde,
-  0x18, 0xc3, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x2c, 0xff, 0xff, 0xff, 0xff, 0x7b, 0xef, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xb5, 0xd6,
-  0xff, 0xff, 0xff, 0xff, 0x29, 0x45, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff,
-  0x84, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c,
-  0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x00, 0x20, 0x21, 0x24, 0x21, 0x24, 0x21, 0x24,
-  0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x44, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff,
-  0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00,
-  0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x29, 0x45, 0x00, 0x00, 0xbd, 0xf7,
-  0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff,
-  0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24,
-  0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c,
-  0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff,
-  0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00,
-  0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7,
-  0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff,
-  0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24,
-  0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c,
-  0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff,
-  0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00,
-  0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7,
-  0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff,
-  0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24,
-  0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c,
-  0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff,
-  0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00,
-  0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x29, 0x45, 0x00, 0x00, 0xbd, 0xf7,
-  0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff,
-  0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24,
-  0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c,
-  0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff,
-  0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00,
-  0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7,
-  0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff,
-  0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24,
-  0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c,
-  0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff,
-  0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00,
-  0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7,
-  0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff,
-  0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24,
-  0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c,
-  0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff,
-  0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00,
-  0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7,
-  0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff,
-  0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24,
-  0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c,
-  0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff,
-  0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x5b, 0x0b, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00,
-  0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbd, 0xf7,
-  0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x5a, 0xeb, 0xff, 0xff, 0xff, 0xff,
-  0x84, 0x10, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24,
-  0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x82, 0xce, 0x79,
-  0xff, 0xff, 0xff, 0xff, 0x84, 0x30, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0x21, 0x24, 0x00, 0x00, 0xbe, 0x17, 0xff, 0xff, 0xff, 0xff, 0xa5, 0x34, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x24,
-  0xef, 0x7d, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x73, 0x8e, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x21, 0x44, 0x00, 0x00, 0xad, 0x55, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xc6, 0x58, 0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x18, 0xc3, 0xe7, 0x5c, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x63, 0x2c, 0x00, 0x00, 0x00, 0x00,
-  0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x29, 0x65, 0x00, 0x00, 0x00, 0x00,
-  0x8c, 0x51, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xc6, 0x38, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0xef, 0x7d, 0x29, 0x65, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x21, 0x04, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x29, 0x45,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x5a, 0xcb, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x94, 0xb2,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x0c, 0xff, 0xff, 0xff, 0xff, 0xf7, 0xde, 0x29, 0x45,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x52, 0xaa, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0x5a, 0xcb, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x52, 0x8a, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0x31, 0xa6, 0x00, 0x00, 0x00, 0x00, 0xc6, 0x38, 0xff, 0xff, 0xff, 0xff,
-  0x6b, 0x4d, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0xb5, 0xb6, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xbd, 0xf7, 0x08, 0x41, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x94, 0xb2, 0xff, 0xff, 0xff, 0xff, 0x9d, 0x13, 0x00, 0x00, 0x42, 0x28, 0xff, 0xff,
-  0xff, 0xff, 0xd6, 0xba, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xce, 0x79, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd6, 0xba,
-  0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 0x21, 0x24, 0xf7, 0xde, 0xff, 0xff, 0xf7, 0xbe, 0x18, 0xe3,
-  0x7b, 0xef, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x8c, 0x51,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0x8c, 0x91, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xa5, 0x34, 0xff, 0xff,
-  0xff, 0xff, 0x5a, 0xeb, 0xbd, 0xf7, 0xff, 0xff, 0xff, 0xff, 0x31, 0xc6, 0x00, 0x00, 0x00, 0x00,
-  0x10, 0xa2, 0xf7, 0xbe, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf7, 0xde, 0x18, 0xc3, 0x00, 0x00, 0x00, 0x00,
-  0x6b, 0x8d, 0xff, 0xff, 0xff, 0xff, 0x84, 0x10, 0xdf, 0x1b, 0xff, 0xff, 0xef, 0x7d, 0x08, 0x41,
-  0x00, 0x00, 0x00, 0x00, 0x39, 0xe7, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x42, 0x28,
-  0x00, 0x00, 0x00, 0x00, 0x39, 0xe7, 0xff, 0xff, 0xff, 0xff, 0xa5, 0x54, 0xef, 0x7d, 0xff, 0xff,
-  0xe7, 0x5c, 0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 0x52, 0x8a, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0x5a, 0xeb, 0x00, 0x00, 0x00, 0x00, 0x31, 0xa6, 0xff, 0xff, 0xff, 0xff, 0xb5, 0xb6,
-  0xde, 0xfb, 0xff, 0xff, 0xef, 0x7d, 0x08, 0x61, 0x00, 0x00, 0x00, 0x00, 0x42, 0x28, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x4a, 0x49, 0x00, 0x00, 0x00, 0x00, 0x42, 0x08, 0xff, 0xff,
-  0xff, 0xff, 0xa5, 0x34, 0xb5, 0xb6, 0xff, 0xff, 0xff, 0xff, 0x3a, 0x07, 0x00, 0x00, 0x00, 0x00,
-  0x18, 0xc3, 0xf7, 0xbe, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf7, 0xde, 0x18, 0xe3, 0x00, 0x00, 0x00, 0x00,
-  0x73, 0xce, 0xff, 0xff, 0xff, 0xff, 0x7b, 0xef, 0x7b, 0xcf, 0xff, 0xff, 0xff, 0xff, 0x84, 0x50,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x8c, 0x51, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x8c, 0x91, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0xa5, 0x34, 0xff, 0xff, 0xff, 0xff, 0x5a, 0xcb, 0x42, 0x28, 0xff, 0xff,
-  0xff, 0xff, 0xd6, 0xba, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xce, 0x79, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd6, 0xba,
-  0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 0x21, 0x24, 0xff, 0xff, 0xff, 0xff, 0xf7, 0xbe, 0x18, 0xc3,
-  0x00, 0x00, 0xc6, 0x38, 0xff, 0xff, 0xff, 0xff, 0x73, 0xae, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x08, 0x41, 0xb5, 0xb6, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xb5, 0xd6, 0x08, 0x61, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x9c, 0xf3, 0xff, 0xff, 0xff, 0xff,
-  0x9c, 0xf3, 0x00, 0x00, 0x00, 0x00, 0x5a, 0xeb, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x31, 0x86,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x42, 0x08, 0x94, 0x92, 0xbe, 0x17, 0xc6, 0x18,
-  0x94, 0xb2, 0x42, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x5a, 0xeb, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0x31, 0x86, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xb5, 0x96, 0xff, 0xff,
-  0xff, 0xff, 0xe7, 0x5c, 0x21, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x52, 0xaa,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x84, 0x50, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x08, 0x61, 0xdf, 0x1b, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x6b, 0x4d, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x8c, 0x71, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xbd, 0xf7, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x24, 0xef, 0x9d, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xd6, 0xba, 0x63, 0x2c, 0x10, 0x82, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0xe3,
-  0x7b, 0xef, 0xe7, 0x3c, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xce, 0x99, 0x08, 0x41, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0xc3, 0xce, 0x79,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf7, 0xbe, 0xd6, 0xba, 0xc6, 0x58, 0xc6, 0x58,
-  0xd6, 0xda, 0xf7, 0xde, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xad, 0x55, 0x00, 0x20,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x5a, 0xeb, 0xe7, 0x3c, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd6, 0xba, 0x39, 0xe7,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x52, 0xaa, 0xa5, 0x14,
-  0xde, 0xfb, 0xef, 0x7d, 0xf7, 0xde, 0xf7, 0xbe, 0xe7, 0x5c, 0xd6, 0xba, 0x94, 0xb2, 0x42, 0x28,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-};
-#define thermometer_width  26
-#define thermometer_height  64
